requirements-scraper-integration.txt
# Scraper Integration Requirements
# Core dependencies for integrating scraper data with Pinecone RAG

# Firebase Admin SDK
firebase-admin>=6.2.0

# Pinecone Vector Database
pinecone-client>=2.2.4

# OpenAI API Client
openai>=1.3.0

# Environment Variables
python-dotenv>=1.0.0

# Data Processing
PyPDF2>=3.0.0

# Optional: Enhanced PDF Processing
# pdfplumber>=0.9.0  # Alternative PDF text extraction
# pymupdf>=1.23.0   # High-performance PDF processing

# Optional: Data Validation
# pydantic>=2.0.0   # Data validation and settings management

# Optional: Async Support
# aiohttp>=3.8.0    # Async HTTP client for better performance
# asyncio-throttle>=1.0.0  # Rate limiting for API calls
